+++ packages/jetbrains-plugin/src/main/kotlin/com/mongodb/jbplugin/inspections/impl/NamespaceCheckInspectionBridge.kt
@@ -16,27 +16,26 @@ import com.mongodb.jbplugin.i18n.InspectionsAndInlaysMessages
 import com.mongodb.jbplugin.inspections.AbstractMongoDbInspectionBridge
 import com.mongodb.jbplugin.inspections.MongoDbInspection
 import com.mongodb.jbplugin.inspections.quickfixes.OpenConnectionChooserQuickFix
-import com.mongodb.jbplugin.linting.FieldCheckWarning
-import com.mongodb.jbplugin.linting.FieldCheckingLinter
+import com.mongodb.jbplugin.linting.NamespaceCheckWarning
+import com.mongodb.jbplugin.linting.NamespaceCheckingLinter
 import com.mongodb.jbplugin.mql.Node
-import com.mongodb.jbplugin.mql.components.HasCollectionReference
 import kotlinx.coroutines.CoroutineScope
 
 /**
  * @param coroutineScope
  */
 @Suppress("MISSING_KDOC_TOP_LEVEL")
-class FieldCheckInspectionBridge(coroutineScope: CoroutineScope) :
+class NamespaceCheckInspectionBridge(coroutineScope: CoroutineScope) :
     AbstractMongoDbInspectionBridge(
         coroutineScope,
-        FieldCheckLinterInspection,
+        NamespaceCheckingLinterInspection,
     )
 
 /**
  * This inspection object calls the linting engine and transforms the result so they can be rendered in the IntelliJ
  * editor.
  */
-internal object FieldCheckLinterInspection : MongoDbInspection {
+internal object NamespaceCheckingLinterInspection : MongoDbInspection {
     override fun visitMongoDbQuery(
         coroutineScope: CoroutineScope,
         dataSource: LocalDataSource?,
@@ -45,17 +44,13 @@ internal object FieldCheckLinterInspection : MongoDbInspection {
         formatter: DialectFormatter,
     ) {
         if (dataSource == null || !dataSource.isConnected()) {
-            return registerNoConnectionProblem(coroutineScope, problems, query.source)
-        }
-
-        if (query.component<HasCollectionReference<PsiElement>>()?.reference is HasCollectionReference.OnlyCollection) {
-            return registerNoDatabaseSelectedProblem(coroutineScope, problems, query.source)
+            return
         }
 
         val readModelProvider = query.source.project.getService(DataGripBasedReadModelProvider::class.java)
 
         val result =
-            FieldCheckingLinter.lintQuery(
+            NamespaceCheckingLinter.lintQuery(
                 dataSource,
                 readModelProvider,
                 query,
@@ -63,20 +58,23 @@ internal object FieldCheckLinterInspection : MongoDbInspection {
 
         result.warnings.forEach {
             when (it) {
-                is FieldCheckWarning.FieldDoesNotExist -> registerFieldDoesNotExistProblem(coroutineScope, problems, it)
-                is FieldCheckWarning.FieldValueTypeMismatch ->
-                    registerFieldValueTypeMismatch(coroutineScope, problems, it, formatter)
+                is NamespaceCheckWarning.NoNamespaceInferred ->
+                    registerNoNamespaceInferred(coroutineScope, problems, it.source)
+                is NamespaceCheckWarning.CollectionDoesNotExist ->
+                    registerCollectionDoesNotExist(coroutineScope, problems, it.source, it.database, it.collection)
+                is NamespaceCheckWarning.DatabaseDoesNotExist ->
+                    registerDatabaseDoesNotExist(coroutineScope, problems, it.source, it.database)
             }
         }
     }
 
-    private fun registerNoConnectionProblem(
+    private fun registerNoNamespaceInferred(
         coroutineScope: CoroutineScope,
         problems: ProblemsHolder,
         source: PsiElement
     ) {
         val problemDescription = InspectionsAndInlaysMessages.message(
-            "inspection.field.checking.error.message.no.connection",
+            "inspection.namespace.checking.error.message",
         )
         problems.registerProblem(
             source,
@@ -91,13 +89,15 @@ internal object FieldCheckLinterInspection : MongoDbInspection {
         )
     }
 
-    private fun registerNoDatabaseSelectedProblem(
+    private fun registerDatabaseDoesNotExist(
         coroutineScope: CoroutineScope,
         problems: ProblemsHolder,
-        source: PsiElement
+        source: PsiElement,
+        dbName: String,
     ) {
         val problemDescription = InspectionsAndInlaysMessages.message(
-            "inspection.field.checking.error.message.no.database",
+            "inspection.namespace.checking.error.message.database.missing",
+            dbName
         )
         problems.registerProblem(
             source,
@@ -112,46 +112,28 @@ internal object FieldCheckLinterInspection : MongoDbInspection {
         )
     }
 
-    private fun registerFieldDoesNotExistProblem(
+    private fun registerCollectionDoesNotExist(
         coroutineScope: CoroutineScope,
         problems: ProblemsHolder,
-        warningInfo: FieldCheckWarning.FieldDoesNotExist<PsiElement>,
+        source: PsiElement,
+        dbName: String,
+        collName: String
     ) {
         val problemDescription = InspectionsAndInlaysMessages.message(
-            "inspection.field.checking.error.message",
-            warningInfo.field,
-            warningInfo.namespace,
+            "inspection.namespace.checking.error.message.collection.missing",
+            collName,
+            dbName
         )
-        problems.registerProblem(
-            warningInfo.source,
-            problemDescription,
-            ProblemHighlightType.WARNING,
-            OpenConnectionChooserQuickFix(
-                coroutineScope,
-                InspectionsAndInlaysMessages.message("inspection.field.checking.quickfix.choose.new.connection"),
-            ),
-        )
-    }
 
-    private fun registerFieldValueTypeMismatch(
-        coroutineScope: CoroutineScope,
-        problems: ProblemsHolder,
-        warningInfo: FieldCheckWarning.FieldValueTypeMismatch<PsiElement>,
-        formatter: DialectFormatter,
-    ) {
-        val problemDescription = InspectionsAndInlaysMessages.message(
-            "inspection.field.checking.error.message.value.type.mismatch",
-            formatter.formatType(warningInfo.valueType),
-            formatter.formatType(warningInfo.fieldType),
-            warningInfo.field,
-        )
         problems.registerProblem(
-            warningInfo.valueSource,
+            source,
             problemDescription,
             ProblemHighlightType.WARNING,
             OpenConnectionChooserQuickFix(
                 coroutineScope,
-                InspectionsAndInlaysMessages.message("inspection.field.checking.quickfix.choose.new.connection"),
+                InspectionsAndInlaysMessages.message(
+                    "inspection.field.checking.quickfix.choose.new.connection"
+                ),
             ),
         )
     }