+++ packages/mongodb-mql-model/src/test/kotlin/com/mongodb/jbplugin/mql/components/HasCollectionReferenceTest.kt
@@ -1,28 +1,37 @@
 package com.mongodb.jbplugin.mql.components
 
 import com.mongodb.jbplugin.mql.Namespace
-import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 
 class HasCollectionReferenceTest {
     @Test
     fun `when the underlying reference is Known, it creates a copy with the database modified`() {
+        val dbRef = 1
+        val collRef = 2
+
         val collectionReference = HasCollectionReference(
             HasCollectionReference.Known(
+                dbRef,
+                collRef,
                 Namespace("foo", "bar"),
             )
         )
 
         val modifiedReference = collectionReference.copy("goo")
         // original is not modified
-        assertEquals((collectionReference.reference as HasCollectionReference.Known).namespace.database, "foo")
+        assertEquals((collectionReference.reference as HasCollectionReference.Known).databaseSource, dbRef)
+        assertEquals((collectionReference.reference as HasCollectionReference.Known).collectionSource, collRef)
         assertEquals((modifiedReference.reference as HasCollectionReference.Known).namespace.database, "goo")
     }
 
     @Test
     fun `when the underlying reference is OnlyCollection, it converts it to Known`() {
+        val collRef = 1
         val collectionReference = HasCollectionReference(
             HasCollectionReference.OnlyCollection(
+                collRef,
                 "bar",
             )
         )
@@ -33,6 +42,7 @@ class HasCollectionReferenceTest {
 
         assertTrue(modifiedReference.reference is HasCollectionReference.Known)
         assertEquals((modifiedReference.reference as HasCollectionReference.Known).namespace.database, "foo")
+        assertEquals((modifiedReference.reference as HasCollectionReference.Known).collectionSource, collRef)
     }
 
     @Test